requirements.txt
atomicwrites==1.4.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
oauthlib==3.1.0
pandas==1.3.4
py==1.10.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.3.22
threadpoolctl==3.0.0
toml==0.10.2
tweepy==3.10.0
Werkzeug==2.0.2
wincertstore==0.2
